+++ test/test_helper.rb
@@ -10,9 +10,12 @@ require "action_view/testing/resolvers"
 require "active_model"
 
 ActionView::TestCase.include(Rails::Dom::Testing::Assertions)
+ActionView::Base.field_error_proc = proc { |html_tag, _instance| html_tag }
 
 class User
   include ActiveModel::Model
 
   attr_accessor :name
+
+  validates :name, presence: true
 end